+++ db/c_test.c
@@ -2068,6 +2068,9 @@ int main(int argc, char** argv) {
     rocksdb_options_set_blob_file_starting_level(o, 5);
     CheckCondition(5 == rocksdb_options_get_blob_file_starting_level(o));
 
+    rocksdb_options_set_prepopulate_blob_cache(o, 1 /* flush only */);
+    CheckCondition(1 == rocksdb_options_get_prepopulate_blob_cache(o));
+
     // Create a copy that should be equal to the original.
     rocksdb_options_t* copy;
     copy = rocksdb_options_create_copy(o);