+++ db_stress_tool/db_stress_gflags.cc
@@ -474,6 +474,10 @@ DEFINE_int32(blob_cache_numshardbits, 6,
              "the block and blob caches are different "
              "(use_shared_block_and_blob_cache = false).");
 
+DEFINE_int32(prepopulate_blob_cache, 0,
+             "[Integrated BlobDB] Pre-populate hot/warm blobs in blob cache. 0 "
+             "to disable and 1 to insert during flush.");
+
 static const bool FLAGS_subcompactions_dummy __attribute__((__unused__)) =
     RegisterFlagValidator(&FLAGS_subcompactions, &ValidateUint32Range);