+++ db_stress_tool/db_stress_test_base.cc
@@ -270,6 +270,8 @@ bool StressTest::BuildOptionsTable() {
                         std::vector<std::string>{"0", "1M", "4M"});
     options_tbl.emplace("blob_file_starting_level",
                         std::vector<std::string>{"0", "1", "2"});
+    options_tbl.emplace("prepopulate_blob_cache",
+                        std::vector<std::string>{"kDisable", "kFlushOnly"});
   }
 
   options_table_ = std::move(options_tbl);
@@ -2401,9 +2403,12 @@ void StressTest::Open(SharedState* shared) {
     fprintf(stdout,
             "Integrated BlobDB: blob cache enabled, block and blob caches "
             "shared: %d, blob cache size %" PRIu64
-            ", blob cache num shard bits: %d\n",
+            ", blob cache num shard bits: %d, blob cache prepopulated: %s\n",
             FLAGS_use_shared_block_and_blob_cache, FLAGS_blob_cache_size,
-            FLAGS_blob_cache_numshardbits);
+            FLAGS_blob_cache_numshardbits,
+            options_.prepopulate_blob_cache == PrepopulateBlobCache::kFlushOnly
+                ? "flush only"
+                : "disable");
   } else {
     fprintf(stdout, "Integrated BlobDB: blob cache disabled\n");
   }
@@ -3042,6 +3047,17 @@ void InitializeOptionsFromFlags(
                 "<= 0.\n");
         exit(1);
       }
+    }
+    switch (FLAGS_prepopulate_blob_cache) {
+      case 0:
+        options.prepopulate_blob_cache = PrepopulateBlobCache::kDisable;
+        break;
+      case 1:
+        options.prepopulate_blob_cache = PrepopulateBlobCache::kFlushOnly;
+        break;
+      default:
+        fprintf(stderr, "Unknown prepopulate blob cache mode\n");
+        exit(1);
     }
   }